fix max overlay

+++ tomosaic/register/morph.py
@@ -97,11 +97,11 @@ def arrange_image(img1, img2, shift, order=1, trim=True):
     newimg : ndarray
         Output array.
     """
-    rough_shift = get_roughshift(shift)
+    rough_shift = get_roughshift(shift).astype('int')
     adj_shift = shift - rough_shift.astype('float')
     img2 = realign_image(img2, adj_shift)
     if trim:
-        temp = np.zeros(img2.shape-np.ceil(np.abs(adj_shift)))
+        temp = np.zeros(img2.shape-np.ceil(np.abs(adj_shift)).astype('int'))
         temp[:, :] = img2[:temp.shape[0], :temp.shape[1]]
         img2 = np.copy(temp)
         temp = 0
